Handle new u-boot location
The secure u-boot is no longer stored as a file in the boot partition.
For this reason, the diff mechanism has to be made aware of the way to
extract and compare u-boot binaries in images.
This commit does that, and provides the new u-boot as a regular file in
the boot tarball as needed.

Signed-off-by: Cristian Prundeanu <cristian.prundeanu@arm.com>